myfistsite/pyproject.toml
[tool.poetry]
name = "myfistsite"
version = "0.1.0"
description = "Учебный Django-проект: магазин + аккаунты + блог + API docs + monitoring"
authors = ["OpenAI <noreply@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
Django = "6.0.4"
Pillow = "12.2.0"
djangorestframework = "*"
django-filter = "*"
drf-spectacular = "*"
django-debug-toolbar = "6.3.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
